Rewrote VisualVideo package to use OSMFfor video playback
This replaces the previous use of a simpler
NetStream/NetConnection-based package. This is done mostly to be able
to support different kinds of streams at a later stage, but it also
vastly simplifies the package code.

There are major drawbacks to this approach. Mainly, OSMF adds to the
footprint of the resulting file (188 KB vs about 5 KB before) and it
also introduces some performance problems to be addressed soon. For
example, the current version suffers for issues with audio sync.